feat(sedes): UI para Editar Sede (modal reutilizado y poblado con datos)
