--- src/app/(main)/admin/sedes/columns.tsx
@@ -1,121 +1,0 @@
-"use client";
-
-import { ColumnDef } from "@tanstack/react-table";
-import { ArrowUpDown, MoreHorizontal, Pencil, EyeOff, Eye } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Badge } from "@/components/ui/badge";
-import type { Sede } from '@/lib/validators/sede';
-
-export const columns: ColumnDef<Sede>[] = [
-  {
-    accessorKey: "nombre",
-    header: ({ column }) => (
-      <Button
-        variant="ghost"
-        onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-      >
-        Nombre
-        <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => <div className="font-medium">{row.getValue("nombre")}</div>,
-  },
-  {
-    accessorKey: "direccion",
-    header: "Dirección",
-    cell: ({ row }) => {
-      const direccion = row.getValue("direccion") as string | null;
-      return direccion || <span className="text-xs text-muted-foreground">N/A</span>;
-    },
-  },
-  {
-    accessorKey: "estado",
-    header: ({ column }) => (
-      <Button
-        variant="ghost"
-        onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-      >
-        Estado
-        <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => {
-      const estado = row.getValue("estado") as Sede['estado']; // 'ACTIVO' | 'INACTIVO'
-      const variant = estado === "ACTIVO" ? "default" : "secondary"; // O "destructive" para INACTIVO
-      return <Badge variant={variant} className="capitalize">{estado?.toLowerCase()}</Badge>;
-    },
-    filterFn: (row, id, value) => value.includes(row.getValue(id)),
-  },
-  {
-    accessorKey: "updatedAt",
-    header: ({ column }) => (
-      <Button
-        variant="ghost"
-        onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-      >
-        Última Modificación
-        <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => {
-      const date = new Date(row.getValue("updatedAt"));
-      return <div className="text-sm text-muted-foreground">{date.toLocaleDateString()} {date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}</div>;
-    },
-  },
-  {
-    id: "actions",
-    header: () => <div className="text-right">Acciones</div>,
-    cell: ({ row }) => {
-      const sede = row.original;
-      return (
-        <div className="text-right font-medium">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="ghost" className="h-8 w-8 p-0">
-                <span className="sr-only">Abrir menú</span>
-                <MoreHorizontal className="h-4 w-4" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end">
-              <DropdownMenuLabel>Acciones</DropdownMenuLabel>
-              <DropdownMenuItem
-                // onClick={() => handleEdit(sede)} // Se implementará después
-              >
-                <Pencil className="mr-2 h-4 w-4" />
-                Editar
-              </DropdownMenuItem>
-              <DropdownMenuSeparator />
-              {sede.estado === "ACTIVO" ? (
-                <DropdownMenuItem
-                  className="text-red-600 hover:!text-red-600 dark:hover:!text-red-500 focus:text-red-600 dark:focus:text-red-500"
-                  // onClick={() => handleDeactivate(sede)} // Se implementará después
-                >
-                  <EyeOff className="mr-2 h-4 w-4" />
-                  Desactivar
-                </DropdownMenuItem>
-              ) : (
-                <DropdownMenuItem
-                  className="text-green-600 hover:!text-green-600 dark:hover:!text-green-500 focus:text-green-600 dark:focus:text-green-500"
-                 // onClick={() => handleActivate(sede)} // Se implementará después
-                >
-                  <Eye className="mr-2 h-4 w-4" />
-                  Activar
-                </DropdownMenuItem>
-              )}
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
-      );
-    },
-    enableSorting: false,
-    enableHiding: false,
-  },
-];
