+++ src/app/(main)/admin/sedes/sedes-page-client-content.tsx
@@ -2,12 +2,20 @@
 
 import React from 'react';
 import { Button } from "@/components/ui/button";
-import { PlusCircle, Terminal } from "lucide-react";
+import { PlusCircle, Terminal, MoreHorizontal, Pencil, EyeOff, Eye, ArrowUpDown } from "lucide-react"; // Iconos añadidos
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
+import { Badge } from "@/components/ui/badge"; // Para estado en columnas
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import type { ColumnDef } from "@tanstack/react-table"; // Para definir columnas
 
 import type { Sede } from '@/lib/validators/sede';
-
-import { columns } from './columns';
 import { SedesDataTable } from './sedes-data-table';
 import { SedeFormDialog } from './sede-form-dialog';
 
@@ -20,15 +28,119 @@ export function SedesPageClientContent({
   initialSedes,
   initialFetchError,
 }: SedesPageClientContentProps) {
-  const [isCreateDialogOpen, setIsCreateDialogOpen] = React.useState(false);
+  const [isDialogOpen, setIsDialogOpen] = React.useState(false);
+  const [editingSede, setEditingSede] = React.useState<Sede | null>(null); 
 
-  // Lógica para manejar el resultado de la creación/edición se añadirá aquí
-  // y se pasará a SedeFormDialog.
+  const handleOpenCreateDialog = () => {
+    setEditingSede(null);
+    setIsDialogOpen(true);
+  };
+
+  const handleOpenEditDialog = (sedeToEdit: Sede) => {
+    setEditingSede(sedeToEdit);
+    setIsDialogOpen(true);
+  };
+
+  const handleCloseDialog = () => {
+    setIsDialogOpen(false);
+  };
+
+  // Definimos las columnas aquí para tener acceso a handleOpenEditDialog
+  const tableColumns: ColumnDef<Sede>[] = [
+    {
+      accessorKey: "nombre",
+      header: ({ column }) => (
+        <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
+          Nombre <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      ),
+      cell: ({ row }) => <div className="font-medium">{row.getValue("nombre")}</div>,
+    },
+    {
+      accessorKey: "direccion",
+      header: "Dirección",
+      cell: ({ row }) => {
+        const direccion = row.getValue("direccion") as string | null;
+        return direccion || <span className="text-xs text-muted-foreground">N/A</span>;
+      },
+    },
+    {
+      accessorKey: "estado",
+      header: ({ column }) => (
+        <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
+          Estado <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      ),
+      cell: ({ row }) => {
+        const estado = row.getValue("estado") as Sede['estado'];
+        const variant = estado === "ACTIVO" ? "default" : "secondary";
+        return <Badge variant={variant} className="capitalize">{estado?.toLowerCase()}</Badge>;
+      },
+    },
+    {
+      accessorKey: "updatedAt",
+      header: ({ column }) => (
+        <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
+          Última Modificación <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      ),
+      cell: ({ row }) => {
+        const date = new Date(row.getValue("updatedAt"));
+        return <div className="text-sm text-muted-foreground">{date.toLocaleDateString()} {date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}</div>;
+      },
+    },
+    {
+      id: "actions",
+      header: () => <div className="text-right">Acciones</div>,
+      cell: ({ row }) => {
+        const sede = row.original;
+        return (
+          <div className="text-right font-medium">
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
+                <Button variant="ghost" className="h-8 w-8 p-0">
+                  <span className="sr-only">Abrir menú</span>
+                  <MoreHorizontal className="h-4 w-4" />
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent align="end">
+                <DropdownMenuLabel>Acciones</DropdownMenuLabel>
+                <DropdownMenuItem onClick={() => handleOpenEditDialog(sede)}>
+                  <Pencil className="mr-2 h-4 w-4" />
+                  Editar
+                </DropdownMenuItem>
+                <DropdownMenuSeparator />
+                {sede.estado === "ACTIVO" ? (
+                  <DropdownMenuItem
+                    className="text-red-600 hover:!text-red-600 dark:hover:!text-red-500 focus:text-red-600 dark:focus:text-red-500"
+                    // onClick={() => handleDeactivate(sede)} // Se implementará después
+                  >
+                    <EyeOff className="mr-2 h-4 w-4" />
+                    Desactivar
+                  </DropdownMenuItem>
+                ) : (
+                  <DropdownMenuItem
+                    className="text-green-600 hover:!text-green-600 dark:hover:!text-green-500 focus:text-green-600 dark:focus:text-green-500"
+                   // onClick={() => handleActivate(sede)} // Se implementará después
+                  >
+                    <Eye className="mr-2 h-4 w-4" />
+                    Activar
+                  </DropdownMenuItem>
+                )}
+              </DropdownMenuContent>
+            </DropdownMenu>
+          </div>
+        );
+      },
+      enableSorting: false,
+      enableHiding: false,
+    },
+  ];
 
   return (
     <>
       <div className="flex justify-end mb-6">
-        <Button onClick={() => setIsCreateDialogOpen(true)}>
+        <Button onClick={handleOpenCreateDialog}> 
           <PlusCircle className="mr-2 h-4 w-4" />
           Crear Nueva Sede
         </Button>
@@ -43,17 +155,16 @@ export function SedesPageClientContent({
       )}
 
       <SedesDataTable 
-        columns={columns} 
+        columns={tableColumns} // Usa las columnas definidas localmente
         data={initialSedes} 
         searchColumnId="nombre" 
         searchPlaceholder="Filtrar por nombre de sede..."
       />
 
-      {/* El diálogo para crear una sede */}
       <SedeFormDialog 
-        open={isCreateDialogOpen}
-        onOpenChange={setIsCreateDialogOpen}
-        // La prop onSubmitAction se conectará en el siguiente commit
+        open={isDialogOpen}
+        onOpenChange={handleCloseDialog}
+        initialData={editingSede} // Pasa la sede a editar (o null si es creación)
       />
     </>
   );
